+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Change Log
 
 ## 2.1.3 (TBD)
 
+### Added:
+- Option to represent Array as ArrayList.
+
 ### Fixed
 - Data type for type alias pointing to Maps and Arrays.